Separated the e2e test cases into individual files. (#1524)
